Start function ph (#13)
* npm i changed package-lock.json

* begin setup wizard functionality

* github workflows yaml

* github workflows yaml

* regex for time format testing

* add newChannel function to setup wizard

* revise file name and move logic for start command
